refactor: item page and text as component
