+++ src/mockData.js
@@ -1,6 +1,11 @@
 import Mock from "mockjs";
 const brands = ["LENOVO", "ACER", "微星", "華碩", "雷蛇", "DELL", "技嘉", "威剛", "三星"];
 
+Mock.Random.extend({
+  brand: function() {
+    return this.pick(brands);
+  }
+});
 const mockData = Mock.mock({
   "categories|1-5": [
     {
@@ -15,16 +20,16 @@ const mockData = Mock.mock({
           "imgSrc|1": ["src/assets/cool-1.jpeg", "src/assets/item-6.jpeg"],
           price: "@integer(100, 30000)",
           isHotItem: "@boolean",
-          size: "@sentence(1, 5)",
-          cpu: "@sentence(1, 2)",
-          ram: "@sentence(1, 3)",
-          ssd: "@sentence(1, 2)",
-          vga: "@sentence(1, 4)",
-          lan: "@sentence(1, 3)",
-          other: "@string(1, 2)",
-          os: "@sentence(1, 3)",
-          brand: "@pick(brands)"
-        },
+            size: "@sentence(1, 5)",
+            cpu: "@sentence(1, 2)",
+            ram: "@sentence(1, 3)",
+            ssd: "@sentence(1, 2)",
+            vga: "@sentence(1, 4)",
+            lan: "@sentence(1, 3)",
+            other: "@string(1, 2)",
+            os: "@sentence(1, 3)",
+            brand: "@brand"
+          }
       ],
     },
   ],
